Added a controller to the backend structure
