+++ back-end/server.js
@@ -14,7 +14,7 @@ const PORT = 5000 || process.env.PORT;
 
 connectDB();
 
-app.use('/user', require('./api/routes/users'));
+app.use('/user', require('./api/routes/userRouters'));
 
 app.listen(PORT, () => {
     console.log("Server started on port ${PORT}");
